functional contracts working

+++ src/Models/ExpectedNodeState.cs
@@ -1,3 +1,5 @@
+using System.Numerics;
+
 namespace src
 {
     public class ExpectedNodeState
@@ -7,6 +9,6 @@ namespace src
         public string ChainspecUrl { get; set; }
         public string ChainspecChecksum { get; set; }
         public bool IsSigning { get; set; }
-        public long UpdateIntroducedBlock { get; set; }
+        public BigInteger UpdateIntroducedBlock { get; set; }
     }
 }
